Ressurect QuicFizzFactory
